slopsmith/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
pycryptodome>=3.15.0
pyguitarpro>=0.10.1
cloudscraper>=1.2.71
curl_cffi>=0.7.0
midiutil>=1.2.1
mido>=1.3.0
Pillow>=10.0.0
PyYAML>=6.0
python-multipart>=0.0.9
structlog>=24.1.0
asgi-correlation-id>=4.3.3
psutil>=5.9.0
py-cpuinfo>=9.0.0
